+++ src/main/java/dnd/myOcean/service/member/MemberService.java
@@ -1,48 +1,9 @@
 package dnd.myOcean.service.member;
 
-import dnd.myOcean.domain.member.Member;
-import dnd.myOcean.dto.oAuth.response.OAuthAttributes;
-import dnd.myOcean.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.oauth2.client.userinfo.DefaultOAuth2UserService;
-import org.springframework.security.oauth2.client.userinfo.OAuth2UserRequest;
-import org.springframework.security.oauth2.client.userinfo.OAuth2UserService;
-import org.springframework.security.oauth2.core.OAuth2AuthenticationException;
-import org.springframework.security.oauth2.core.user.DefaultOAuth2User;
-import org.springframework.security.oauth2.core.user.OAuth2User;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import java.util.Collections;
 
 @Service
 @RequiredArgsConstructor
-public class MemberService implements OAuth2UserService<OAuth2UserRequest, OAuth2User> {
-
-    private final MemberRepository memberRepository;
-    private final OAuth2UserService<OAuth2UserRequest, OAuth2User> delegate = new DefaultOAuth2UserService();
-
-    @Transactional
-    @Override
-    public OAuth2User loadUser(OAuth2UserRequest userRequest) throws OAuth2AuthenticationException {
-        OAuth2User oAuth2User = delegate.loadUser(userRequest);
-
-        OAuthAttributes attributes = OAuthAttributes.of("id", oAuth2User.getAttributes());
-        Member member = saveOrUpdate(attributes);
-
-        return new DefaultOAuth2User(
-                Collections.singleton(new SimpleGrantedAuthority(member.getRoleKey())),
-                attributes.getAttributes(),
-                attributes.getNameAttributeKey()
-        );
-    }
-
-    private Member saveOrUpdate(OAuthAttributes attributes) {
-        Member member = memberRepository.findByEmail(attributes.getEmail())
-                .map(entity -> entity.update(attributes.getName(), attributes.getProvider()))
-                .orElse(attributes.toEntity());
-
-        return memberRepository.save(member);
-    }
+public class MemberService {
 }
